UI/UX Overhaul
Adjusted the dropdowns to be more functional search boxes (also optimizes it for mobile users)

Added styling for the dropdowns

Adjusted the initializeDraggableCards function to work with the new dropdowns and correctly allow for monsters to swap places

Adjusted the base CSS so that the boxes all stay on the same plane and all of the elements are fixed.

Added a small note for mobile users in the header.