Claire fix
Adjusted Claire skills to remove the double ATB Boost effect. One was set to Self_effect true and the other was self_effect false. It kept the latter one. It was removed to prevent Claire ATB Boost from showing.

Added console logs to check for this in the future for bug checks.

Removed some empty space in the code.